requirements.txt
psycopg2-binary
requests
faker
matplotlib
python-dotenv
reportlab
